requirements.txt
astroid==2.3.3
async==0.6.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
future==0.18.2
get==2019.4.13
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
mccabe==0.6.1
post==2019.4.13
public==2019.4.13
pylint==2.4.4
python-firebase==1.2
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
